# expensetrackerapi/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Expense Tracker",
    "description": "Expense Tracker",
    "version": "1.0.0"
  },
  "host": "localhost:1337",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/user/save": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "User save service",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Request body.",
            "schema": {
              "$ref": "#/definitions/userSave"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User created successfully."
          }
        }
      }
    },
...
    "/api/user/resetpassword": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "userName",
            "in": "header",
            "type": "string"
          },
          {
            "name": "newpassword",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "request": {
                  "example": "any"
                },
                "newpassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}